refactor: 문자열 처리 로직을 StringProcessor로 분리
